==> 3.EDA_clustering.ipynb <==
# 3. EDA_clustering

Try to build clusters to find patterns before labeling.  
Results is not good as expected-- we won't use unsupervised learning in our project.

## Using clusters to find patterns

# read data
%store -r disaster_tweets
%store -r tweets_words

tweets_words.head()

disaster_tweets.head()

len(tweets_words)

len(disaster_tweets)

## k-means

We chose k-means here because it can control the number of clusters.

from sklearn.cluster import KMeans
from sklearn.metrics import silhouette_score
from sklearn.preprocessing import StandardScaler

sc = StandardScaler()
X = sc.fit_transform(tweets_words)

km = KMeans(n_clusters=2, #we tried 2, 3 and 5 to see if there are any patterns
            random_state=42)
km.fit(X)

len(km.labels_)

len(disaster_tweets)

disaster_tweets['cluster'] = km.labels_
disaster_tweets.head()

# number of cluster 0
len(disaster_tweets[disaster_tweets['cluster'] == 0]['text'])

# number of cluster 1
len(disaster_tweets[disaster_tweets['cluster'] == 1]['text'])
