==> 4.modeling.ipynb <==
# 4.modeling

Label tweets as 0 and 1 based on key words to train classification models.

## 4.1 Label tweets based on key words

# read data
%store -r disaster_tweets
%store -r self_defined_stop_words
%store -r final_disaster_words

import pandas as pd
import numpy as np
from sklearn.ensemble import VotingClassifier, RandomForestClassifier, ExtraTreesClassifier, BaggingClassifier, AdaBoostClassifier, GradientBoostingClassifier
from sklearn.linear_model import LinearRegression, LassoCV, RidgeCV, LogisticRegression
from sklearn.tree import DecisionTreeClassifier, DecisionTreeRegressor
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.neighbors import KNeighborsRegressor, KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.metrics import confusion_matrix
from sklearn.feature_extraction import stop_words 
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import MultinomialNB, ComplementNB
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer 
from sklearn.metrics import mean_squared_error, f1_score, confusion_matrix
from imblearn.over_sampling import RandomOverSampler
from collections import Counter


disaster_tweets.reset_index(inplace=True, drop=True)

# we tried self defined words to label tweets as well
# but our final_disaster_words list works better

# words=[
#  'need shelter',
#  'help me',
#  'needhelp',
#  'help us',
#  'pleasehelp',
#  'casualties',
#  'assistance',
#  'need help',
#  'send help',
#  'please help',
#  'shelter',
#  'emergency']

words = final_disaster_words

disaster_tweet_index = []
words = final_disaster_words

for i in words:
    for x in range(len(disaster_tweets['text'])):
        if i in disaster_tweets['text'][x]:
            disaster_tweet_index.append(x)
print(f'Number of tweets related to disaster is {len(disaster_tweet_index)}')
print(f'Number of overall tweets is {len(disaster_tweets)}')


# label tweets as 0 or 1
disaster_tweets['label'] = 0
disaster_tweets['label'][disaster_tweet_index] = 1

## 4.2 Build models

X = disaster_tweets['text']
y = disaster_tweets['label']

# resampling to create a balanced class
ros = RandomOverSampler(random_state=0) # Instantiate a random oversampler in order to oversample our training set
X_resampled, y_resampled = ros.fit_resample(X.values.reshape(-1, 1), y) # Fit that oversampler to our X_sc (scaled) and y data
print(sorted(Counter(y_resampled).items())) # Show the balance btw classes

# train test data split in training set
# test data here is used to evaluate
X_train, X_test, y_train, y_test = train_test_split(X_resampled, y_resampled, random_state=42)

# count vectorizer words
class LemmaTokenizer(object):
    def __init__(self):
        self.wnl = WordNetLemmatizer()
    def __call__(self, doc):
        tokenizer = RegexpTokenizer('(?u)\\b\\w\\w+\\b')
        return [self.wnl.lemmatize(t, 'v') for t in tokenizer.tokenize(doc)]

from sklearn.feature_extraction import text 
stop_words = text.ENGLISH_STOP_WORDS.union(self_defined_stop_words)
vectorizer = CountVectorizer(tokenizer = LemmaTokenizer(),
                            preprocessor = None,
                            stop_words = stop_words,
                            max_features = 1500,
                            ngram_range= (1,2),
                            analyzer = 'word', 
                            min_df=3) 
vectorizer.fit(X_train.ravel())
X_train = vectorizer.transform(X_train.ravel())
X_test = vectorizer.transform(X_test.ravel())
X_train_df = pd.DataFrame(X_train.toarray(), columns=vectorizer.get_feature_names())
X_test_df = pd.DataFrame(X_test.toarray(), columns=vectorizer.get_feature_names())
y_train_df = pd.DataFrame(y_train)


# use best params to build classification model
lr_class = LogisticRegression(penalty='l1', C=40, solver='liblinear')
# knn_class = KNeighborsClassifier(n_neighbors=3, p=4, leaf_size=10, metric='minkowski')
tree_class = DecisionTreeClassifier(max_features='auto', min_samples_leaf=3, min_samples_split=4, random_state=100)
bag_class = BaggingClassifier(bootstrap=False, max_features=8, max_samples=100, n_estimators=100, random_state=100)
forest_class = RandomForestClassifier(bootstrap=True, max_leaf_nodes=None, min_samples_leaf=3, min_samples_split=8, n_estimators=9, random_state=100)
ada_class = AdaBoostClassifier(learning_rate=0.78, n_estimators=100, random_state=100)
svc = SVC(degree=8, C=2.5, gamma=0.1, kernel='poly', random_state=100)
grad_class = GradientBoostingClassifier(learning_rate=0.06464371632490062, n_estimators=100, min_samples_leaf=2, min_samples_split=5, max_depth=7)

# use below code to find the best params for each model
# params = {
#     'max_depth' : [5, 6, 7],
#     'min_samples_leaf' : [2, 3],
#     'min_samples_split' : [4, 5, 6],
#     'n_estimators' : [50, 100, 125],
#     "learning_rate" : (np.logspace(-1.6, -1, 20))
# }
# gs = GridSearchCV(
#     GradientBoostingClassifier(),
#     params,
#     cv=3,
#     verbose=1,
#     return_train_score=False,
#     n_jobs=2)
# gs.fit(X_train, y_train)
# print(gs.best_score_)
# print()
# print(gs.best_params_)
# print()
# print(gs.score(X_test, y_test))
# pred = model.predict(X_test)
# f1_score(y_test, pred)

class_models = {
            'lr_class': lr_class, 
            'forest_class': forest_class, 
            'tree_class': tree_class,
            'ada_class': ada_class,            
            'bag_class': bag_class, 
            'svc': svc,
            "grad": grad_class
                } 

y_pred_testc = []
y_pred_trainc = []

for model in class_models.values():
    model.fit(X_train, y_train)
    y_pred_testc.append(model.predict(X_test))
    y_pred_trainc.append(model.predict(X_train))

y_pred_testc_df = pd.DataFrame(y_pred_testc, index=class_models.keys()).T
y_pred_trainc_df = pd.DataFrame(y_pred_trainc, index=class_models.keys()).T
print(y_pred_testc_df.shape)
print(y_pred_trainc_df.shape)

# create summary chart
accuracy = {'train': [], 'test': [], 'F1-train': [], 'F1-test': [], 'true_neg': [], 'fal_pos': [], 'fal_neg': [], 'true_po': []}
for model in class_models.values():
    accuracy['train'].append(model.score(X_train, y_train))
    accuracy['test'].append(model.score(X_test, y_test))
for col in y_pred_testc_df:
    accuracy['F1-train'].append(f1_score(y_train, y_pred_trainc_df[col]))
    accuracy['F1-test'].append(f1_score(y_test, y_pred_testc_df[col]))
for col in y_pred_testc_df:
    accuracy['true_neg'].append(confusion_matrix(y_test, y_pred_testc_df[col])[0][0])
    accuracy['fal_pos'].append(confusion_matrix(y_test, y_pred_testc_df[col])[0][1])
    accuracy['fal_neg'].append(confusion_matrix(y_test, y_pred_testc_df[col])[1][0])
    accuracy['true_po'].append(confusion_matrix(y_test, y_pred_testc_df[col])[1][1])
    
accuracy_df = pd.DataFrame(accuracy, index=class_models.keys())
accuracy_df

## 4.3 Apply models on new tweets

# read new tweets
campfire_new_tweets = pd.read_csv('./data/campfire_new_tweets.csv')
campfire_new_tweets_2 = pd.read_csv('./data/campfire_new_tweets_2.csv')
carrfire_new_tweets = pd.read_csv('./data/carrfire_new_tweets.csv')

# combine results
new_tweets = pd.concat([campfire_new_tweets,campfire_new_tweets_2,carrfire_new_tweets])

# remove duplicate tweets
new_tweets.drop_duplicates(subset ="text", keep = False, inplace = True)
new_tweets.reset_index(inplace=True, drop=True)

# we have 832 tweets to predict
len(new_tweets)

# apply countvectorizer on X_pred
X_pred = new_tweets['text']
X_pred = vectorizer.transform(X_pred.ravel())
X_pred_df = pd.DataFrame(X_pred.toarray(), columns=vectorizer.get_feature_names())


# apply models on X_pred_df
prediction = pd.DataFrame(index=X_pred_df.index)

for (model_name, model) in class_models.items():
    prediction[model_name] = model.predict(X_pred_df)
prediction['total'] = prediction['lr_class'] + prediction['forest_class'] + prediction['tree_class'] + \
        prediction['ada_class'] + prediction['bag_class'] + prediction['svc'] + prediction['grad'] 


# combine all predict together
prediction.sort_values(by='total', ascending=False).head(10)

# number in the array means how many model predicted it as 1
prediction['total'].unique()

# number of tweets marked as 1 by 3 or more models
len(prediction[~((prediction['total']==0)|(prediction['total']==1)|(prediction['total']==2))])

# index of rows that are predicted as "related to disaster and need help"
index_true= prediction[~((prediction['total']==0)|(prediction['total']==1)|(prediction['total']==1))].index

# quick check of related tweets
new_tweets.iloc[index_true]['text'].head(20)

# pull longitute and latitute
index_true= index_true.to_list()
coord_true = new_tweets.iloc[index_true][['latitute','longitute']]

coord_true.head()

# export coordinates as csv file to plot
coord_true.to_csv('./data/coord_to_plot.csv', index=False)
