==> 5.google_maps_plotting.ipynb <==
#Installing gmplot
!pip install gmplot

#Importing relevant Libraries
from gmplot import gmplot
import numpy as np
import pandas as pd

#Creating a dataframe of latitudes and Longitudes
df = pd.read_csv('./data/coord_to_plot.csv')

#Displaying the head of the dataframe
df.head()

#Converting Latitude and Longitude to seperate lists
latitude = df['latitute'].tolist()
longitude = df['longitute'].tolist()

# import gmplot package 
import gmplot 
  
latitude_list = latitude
longitude_list = longitude 
  
gmap3 = gmplot.GoogleMapPlotter(39.822949,-121.41392, 26,
                               apikey = "Your API Key")
  
# scatter method of map object  
# scatter points on the google map 
gmap3.scatter(latitude_list, longitude_list,size = 1000, marker = True, color = 'red' ) 
# Get an output html file of all plots  
gmap3.draw("./image/my_map.html")

# Screenshots were put in image folder
